feature/265 Containerize with docker
This PR makes it possible to run the project with one command

```
pnpm docker-dev
```

Once the update to v14 is done
(https://github.com/shadcn-ui/taxonomy/pull/253)
I would be happy to add testing to it, so we can mock email sending and
test the features of the app.

I hope it is useful. Thanks